Split projection entities into separate files
