+++ Arctic/Sources/Arctic/Projection.swift
@@ -7,30 +7,10 @@
 
 import Foundation
 
-typealias LoanID = String
 
-extension Projection {
-    static func calculate(loans: [Loan], monthlyPaymentAmount: Decimal, strategy: PayoffStrategy = .Avalanche) -> Projection {
-        let oldestPersonEverAgeInYears = 122
-        let reasonablePayoffInMonths = oldestPersonEverAgeInYears * 12
-
-        let startDate = Date()
-        let calendar = Calendar.current
-        var projection = Projection(loans: loans, monthlyPaymentAmount: monthlyPaymentAmount, strategy: strategy)
-        var monthCounter = 0
-        
-        while (projection.hasRemainingLoans) {
-            if (monthCounter > reasonablePayoffInMonths) {
-                break
-            }
-            let date = calendar.date(byAdding: .month, value: monthCounter, to: startDate)!
-            projection.makeMonthlyPayments(date: date)
-
-            monthCounter += 1
-        }
-
-        return projection
-    }
+enum PayoffStrategy {
+    case Snowball
+    case Avalanche
 }
 
 struct Projection {
@@ -39,7 +19,7 @@ struct Projection {
     private var monthlyPaymentAmount: Decimal
     private var remainingPaymentAmount: Decimal
     private let strategy: PayoffStrategy
-
+    
     init(loans: [Loan], monthlyPaymentAmount: Decimal, strategy: PayoffStrategy = .Avalanche) {
         self.mapping = makeLoanPaymentMapping(loans)
         self.loans = loans
@@ -104,7 +84,7 @@ struct Projection {
             return lhs.interestRate > rhs.interestRate
         }
     }
-
+    
     private mutating func removePaidOffLoans() {
         mapping.forEach { (loanID, balanceSheet) in
             if (balanceSheet.isPaidOff) {
@@ -112,11 +92,11 @@ struct Projection {
             }
         }
     }
-
+    
     func findBalanceSheet(id: LoanID) -> LoanAccountBalance {
         return mapping[id]!
     }
-
+    
     var hasRemainingLoans: Bool {
         return !loans.isEmpty
     }
@@ -124,13 +104,13 @@ struct Projection {
     var sortedLoans: [Loan] {
         return loans.sorted(by: payoffStrategySort)
     }
-
+    
     var principalPaidAmount: Decimal {
         return mapping.values.reduce(Decimal(0), { accumulator, sheet in
             accumulator + sheet.loan.startingBalance.amount
         })
     }
-
+    
     var interestPaidAmount: Decimal {
         return mapping.values.reduce(Decimal(0), { accumulator, value in
             return accumulator + value.rows.last!.totalInterestPaid
@@ -138,165 +118,10 @@ struct Projection {
     }
 }
 
-class LoanAccountBalance {
-    let loan: Loan
-    var rows: [AccountBalanceRow]
-    var currentEntry: MonthlyLoanEntry? = nil
-    
-    init(loan: Loan, rows: [AccountBalanceRow]) {
-        self.loan = loan
-        self.rows = rows
-    }
-
-    func startEntry(date: Date) {
-        currentEntry = MonthlyLoanEntry(
-            loan: loan,
-            date: date,
-            interestBalance: lastInterestBalance,
-            principalBalance: lastPrincipalBalance,
-            totalInterestPaid: lastTotalInterestPaid
-        )
-    }
-    
-    func makeMinimumPayment() -> Decimal {
-        return currentEntry!.makeMinimumPayment()
-    }
-    
-    func makeExtraPayment(extraPayment: Decimal) -> Decimal {
-        return currentEntry!.makePrincipalPayment(amount: extraPayment)
-    }
-    
-    func applyInterest() {
-        currentEntry!.applyInterest()
-    }
-    
-    func finishEntry() {
-        guard let safeEntry = currentEntry else { return }
-
-        rows.append(
-            AccountBalanceRow(
-                date: safeEntry.date,
-                loanID: loan.id,
-                interestBalance: safeEntry.interestBalance,
-                principalBalance: safeEntry.principalBalance,
-                totalInterestPaid: safeEntry.totalInterestPaid,
-                paymentAmount: safeEntry.paymentAmount
-            )
-        )
-    }
-    
-    var isPaidOff: Bool {
-        return currentTotalBalance == 0
-    }
-
-    var currentTotalBalance: Decimal {
-        return currentEntry!.interestBalance +  currentEntry!.principalBalance
-    }
-    
-    private var lastTotalInterestPaid: Decimal {
-        return lastEntry?.totalInterestPaid ?? 0
-    }
-
-    private var lastInterestBalance: Decimal {
-        return lastEntry?.interestBalance ?? 0
-    }
-    
-    private var lastPrincipalBalance: Decimal {
-        return lastEntry?.principalBalance ?? loan.startingBalance.amount
-    }
-    
-    private var lastEntry: AccountBalanceRow? {
-        return rows.last
-    }
-}
-
-func makeLoanPaymentMapping(_ loans: [Loan]) -> [LoanID : LoanAccountBalance] {
+fileprivate func makeLoanPaymentMapping(_ loans: [Loan]) -> [LoanID : LoanAccountBalance] {
     var mapping = [LoanID : LoanAccountBalance]()
     loans.forEach { loan in
         mapping.updateValue(LoanAccountBalance(loan: loan, rows: []), forKey: loan.id)
     }
     return mapping
-}
-
-struct AccountBalanceRow {
-    let date: Date
-    let loanID: String
-    let interestBalance: Decimal
-    let principalBalance: Decimal
-    let totalInterestPaid: Decimal
-    let paymentAmount: Decimal
-}
-
-struct MonthlyLoanEntry {
-    let date: Date
-    let loan: Loan
-    var interestBalance: Decimal
-    var principalBalance: Decimal
-    var totalInterestPaid: Decimal
-    var paymentAmount: Decimal = 0
-    
-    init(loan: Loan, date: Date, interestBalance: Decimal, principalBalance: Decimal, totalInterestPaid: Decimal) {
-        self.loan = loan
-        self.date = date
-        self.interestBalance = interestBalance
-        self.principalBalance = principalBalance
-        self.totalInterestPaid = totalInterestPaid
-    }
-    
-    mutating func makeMinimumPayment() -> Decimal {
-        let payment = loan.minimumPayment.amount
-    
-        if (payment > interestBalance) {
-            let paidDown = interestBalance
-            let leftover = payment - paidDown
-            paymentAmount += paidDown
-            totalInterestPaid += paidDown
-            interestBalance = 0
-        
-            return paidDown + makePrincipalPayment(amount: leftover)
-        } else {
-            interestBalance -= payment
-            paymentAmount += payment
-            return payment
-        }
-    }
-    
-    mutating func makePrincipalPayment(amount payment: Decimal) -> Decimal {
-        if (payment > principalBalance) {
-            let paidDown = principalBalance
-            paymentAmount += paidDown
-            principalBalance = 0
-            return paidDown
-        } else {
-            principalBalance -= payment
-            paymentAmount += payment
-            return payment
-        }
-    }
-    
-    mutating func applyInterest() {
-        interestBalance = (currentBalance * interestRatePercent).rounded(2)
-    }
-    
-    var interestRatePercent: Decimal {
-        return loan.interestRate / 100.0 / 12.0
-    }
-    
-    var currentBalance: Decimal {
-        return interestBalance + principalBalance
-    }
-}
-
-enum PayoffStrategy {
-    case Snowball
-    case Avalanche
-}
-
-extension Decimal {
-    func rounded(_ scale: Int = 1) -> Decimal {
-        var decimal = self
-        var rounded = Decimal()
-        NSDecimalRound(&rounded, &decimal, scale, .plain)
-        return rounded
-    }
 }
